move dlgpreferencepage to src/preferences/dialog/
