+++ src/preferences/dialog/dlgprefinterface.h
@@ -3,9 +3,9 @@
 #include <QWidget>
 
 #include "preferences/constants.h"
+#include "preferences/dialog/dlgpreferencepage.h"
 #include "preferences/dialog/ui_dlgprefinterfacedlg.h"
 #include "preferences/usersettings.h"
-#include "preferences/dlgpreferencepage.h"
 
 class ControlProxy;
 class ControlPotmeter;
