+++ src/preferences/dialog/dlgprefkey.h
@@ -7,8 +7,8 @@
 #include "analyzer/plugins/analyzerplugin.h"
 #include "control/controlproxy.h"
 #include "defs_urls.h"
+#include "preferences/dialog/dlgpreferencepage.h"
 #include "preferences/dialog/ui_dlgprefkeydlg.h"
-#include "preferences/dlgpreferencepage.h"
 #include "preferences/keydetectionsettings.h"
 #include "preferences/usersettings.h"
 #include "track/keyutils.h"
